Re-work the compiler to raise errors
A call to Compiler.compile() now holds a Promise instance that it will
return on subsequent calls to .compile() removing the need for the
_running flag.

We now throw compiler errors if WebPage fails to build. This will exit
with a non-zero code during build but be handled by the express error
handler when the server is running.