Add support for building sites with a path prefix
This allows someone to publish their website under a subpath e.g.
/blog and have all the urls work as expected. We currently support
this via relative urls, but these can be fragile depending on how
the host deals with trailing slashes. If the host supports serving
an index.html file at both /about and /about/ any relative urls will
resolve differently.

This commit introduces a new configuration option called `urlPrefix`
that can be provided via the `--prefix` command line flag. We then
use this prefix for internal urls as well as in the web server.

Usage:

    # For the development server:
    $ yarn monobase serve --project=./project --prefix=/api

    # For the build:
    $ yarn monobase build --path=./build/api --prefix=/api

I've left the relative() helper in there for the moment as I can
imagine there are projects depending on it. There should be no change
to projects that do not provide a prefix (components.js will use
relative urls in this case to keep exising projects consistent).